+++ docs/changeLog.md
@@ -13,6 +13,7 @@ order: 1
 
 - FormItem
   - 优化 validatorCallback 使其支持 callback 和 Promise 的返回值
+  - 优化 itemType 使其支持 textarea，内置使用 Input.TextArea
 
 ### 🚀 2.14.0-alpha.8